implementing the stubbed functionality to hook roi merge #279 together, need just a couple typeclass instances in MergeRois and then to update tests
